refactor: Move Pagination.module.scss
