+++ components/pagination/Pagination.module.scss
@@ -1,22 +1,22 @@
-@import './variables';
+@import '@/styles/variables';
 
 .container {
   display: flex;
-  list-style: none;
-  margin: 0;
   padding: 0;
+  margin: 0;
+  list-style: none;
 }
 
 .item {
+  display: flex;
   align-items: center;
+  justify-content: center;
+  width: 32px;
+  height: 32px;
+  margin: 8px;
   border: 1px solid $light-gray;
   border-radius: 10px;
-  display: flex;
-  height: 32px;
-  justify-content: center;
-  margin: 8px;
   transition: color 0.15s ease, border-color 0.15s ease;
-  width: 32px;
 
   span {
     display: inline-block;
@@ -25,8 +25,8 @@
   &:hover,
   &:focus,
   &:active {
-    border-color: $primary-purple;
     color: $primary-purple;
+    border-color: $primary-purple;
   }
 
   &:hover {
@@ -43,6 +43,6 @@
 }
 
 .active {
-  border-color: $primary-purple;
   color: $primary-purple;
+  border-color: $primary-purple;
 }
